+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes after its initial development up to June 2020 (v1.0) are documented in this file.
 
+### v1.4 (Oct 2021)
+    - Move to pyproject.toml structure with Github actions.
+    - Remove History sections from docstrings.
+
 ### v1.3.5 (May 2021)
     - Escape backslash also in params in sub_params_ja routines.